add two layouts and cleanup code (#25)
